% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qens.R
\name{df_to_array}
\alias{df_to_array}
\title{Convert data frames with predictions and observed data to arrays.
It's OK not to include the data.}
\usage{
df_to_array(predictions, y = NULL, model_id_vars, task_id_vars, tau_var, q_var)
}
\arguments{
\item{predictions}{data frame with training set predictions to use for
estimating model weights.}

\item{y}{data frame with observed responses for training set predictions}

\item{model_id_vars}{\code{character} vector naming columns of \code{predictions} that
identify prediction models.}

\item{task_id_vars}{an optional \code{character} vector naming the columns of
\code{predictions} that correspond to task id variables. The default is \code{NULL},
in which case all columns in \code{predictions} other than those specified for
\verb{model_id_vars, }tau_var\verb{, and }q_var` will be used as task id variables.}

\item{tau_var}{string naming a column of \code{predictions} that
identifies the probability level of quantile forecasts. The default is
\code{NULL}, in which case \code{"quantile"} is used if \code{predictions} contains a
column of that name, or otherwise \code{"output_id"} is used if \code{predictions}
contains a column of that name; these are common naming conventions in
infectious disease forecast hubs. If \code{tau_var} is not specified and
\code{predictions} does not contain columns named \code{"quantile"} or \code{"output_id"},
an error is thrown.}

\item{q_var}{string naming a column of \code{predictions} that has values of
predictive quantiles.}
}
\value{
named list with entries:
\itemize{
\item \code{predictions}: array with shape (N, K, M) where N is the number of
prediction tasks, K is the number of quantile levels, and M is the
number of models.
\item \code{y}: array with observed data corresponding to the first dimension of
predictions
\item \code{task_id}: data frame with N rows containing the task id variables
corresponding to the first axis of \code{predictions} and the entries of \code{y}.
Column names are those specified by \code{task_id_vars}
\item \code{tau}: numeric vector of length K with quantile levels corresponding to
the second axis of \code{predictions}.
\item \code{model_id}: data frame with M rows containing the model identifiers
corresponding to the third axis of \code{predictions}. Column names are those
specified by \code{model_id_vars}
}
}
\description{
Convert data frames with predictions and observed data to arrays.
It's OK not to include the data.
}
\details{
This function expands across missing combinations of task id, model,
and quantile level, filling in missing entries in the predictions with
\code{NA_real_}. It also subsets the predictions and the observed data, only
}
