% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_py_dependencies.R
\name{install_py_dependencies}
\alias{install_py_dependencies}
\title{Install python dependencies including \code{tensorflow}, \code{tensorflow_probability},
\code{qenspy}, and the dependencies of those packages. This function simply calls
\code{reticulate::py_install}; see the documentation of that function for more
information about how to set up Python environments.}
\usage{
install_py_dependencies(method = "auto", conda = "auto")
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method is not
available on Windows.}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda-tools]{conda_binary()}} for more details.}
}
\description{
Install python dependencies including \code{tensorflow}, \code{tensorflow_probability},
\code{qenspy}, and the dependencies of those packages. This function simply calls
\code{reticulate::py_install}; see the documentation of that function for more
information about how to set up Python environments.
}
