% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qens.R
\name{new_qens}
\alias{new_qens}
\title{Create a qens object}
\usage{
new_qens(
  param_vec,
  M,
  model_names = rep("", M),
  quantile_levels,
  quantile_groups,
  combine_method,
  loss_trace = NULL
)
}
\arguments{
\item{param_vec}{numeric vector of "raw" parameter estimates}

\item{M}{integer number of models}

\item{model_names}{name for each model}

\item{combine_method}{string specifying combination method;
either "convex_mean" or "convex_median"}

\item{loss_trace}{numeric vector of loss trace per iteration of estimation}

\item{tau}{numeric vector of quantile levels}

\item{tau_grps}{specification of quantile groups}
}
\value{
qens object
}
\description{
Create a qens object
}
