% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qens.R
\name{new_qenspy_from_qens}
\alias{new_qenspy_from_qens}
\title{Instantiate a new Python object representing a quantile ensemble fit from an
R \code{qens} object.}
\usage{
new_qenspy_from_qens(qens_obj)
}
\arguments{
\item{qens_obj}{An S3 object of class \verb{<qens>}}
}
\value{
A pointer to a python \code{qens} object, for use with reticulate
}
\description{
Instantiate a new Python object representing a quantile ensemble fit from an
R \code{qens} object.
}
