% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qens-package.R, R/qens.R
\docType{package}
\name{qens}
\alias{qens}
\alias{qens-package}
\title{qens - Quantile ensembles}
\usage{
qens(
  predictions,
  y,
  model_id_vars = "model",
  task_id_vars = NULL,
  tau_var = NULL,
  q_var = "value",
  combine_method = "mean",
  weighted = FALSE,
  tau_grps = rep(1, length(unique(predictions[[tau_var]]))),
  init_method = if (weighted) "xavier" else "equal",
  optim_method = "adam",
  num_iter = 1000,
  learning_rate = 0.1,
  verbose = FALSE,
  save_frequency = NULL,
  save_filename = NULL
)
}
\arguments{
\item{predictions}{data frame with training set predictions to use for
estimating model weights.}

\item{y}{data frame with observed responses for training set predictions}

\item{model_id_vars}{\code{character} vector naming columns of \code{predictions} that
identify prediction models.}

\item{task_id_vars}{an optional \code{character} vector naming the columns of
\code{predictions} that correspond to task id variables. The default is \code{NULL},
in which case all columns in \code{predictions} other than those specified for
\verb{model_id_vars, }tau_var\verb{, and }q_var` will be used as task id variables.}

\item{tau_var}{string naming a column of \code{predictions} that
identifies the probability level of quantile forecasts. The default is
\code{NULL}, in which case \code{"quantile"} is used if \code{predictions} contains a
column of that name, or otherwise \code{"output_id"} is used if \code{predictions}
contains a column of that name; these are common naming conventions in
infectious disease forecast hubs. If \code{tau_var} is not specified and
\code{predictions} does not contain columns named \code{"quantile"} or \code{"output_id"},
an error is thrown.}

\item{q_var}{string naming a column of \code{predictions} that has values of
predictive quantiles.}

\item{combine_method}{string specifying method for ensembling quantiles;
"mean" or "median"}

\item{weighted}{boolean indicating whether model weighting should be done.
If \code{FALSE}, all models are given equal weight. If \code{TRUE}, model weights are
estimated.}

\item{tau_grps}{Vector of group labels for quantiles, having the same
length as the number of distinct quantile levels in the \code{tau_var} column of
\code{predictions}. The ensemble weights are shared across any quantile levels
with the same label. The default specifies that a common set of weights
should be used across all levels.}

\item{init_method}{string specifying initialization method for parameters.
Choices are "xavier" for random Xavier initialization or "equal" for equal
weights. Ignored if \code{weighted} is \code{FALSE}.}

\item{optim_method}{string specifying optimization method; options are "adam"
and "sgd"}

\item{num_iter}{integer number of iterations for optimization}

\item{learning_rate}{numeric learning rate for optimization}

\item{verbose}{boolean indicating whether to print output during parameter
estimation}
}
\value{
S3 object of class \verb{<qens>}
}
\description{
Quantile ensembles
}
\author{
Evan L. Ray \href{mailto:elray@umass.edu}{elray@umass.edu}
}
