% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qens.R
\name{validate_df_subseteq}
\alias{validate_df_subseteq}
\title{validate that the column names of \code{df_sub} and \code{df_sup} are the same,
and that all rows of df_sub appear in df_sup}
\usage{
validate_df_subseteq(df_sub, df_sup)
}
\arguments{
\item{df_sub}{a data frame that is expected to be a subset}

\item{df_sup}{a dataframe that is expected to be a superset}
}
\value{
TRUE if \code{df_sub} and \code{df_sup} have the same column names and all rows
of df_sub are in df_sup, FALSE otherwise
}
\description{
validate that the column names of \code{df_sub} and \code{df_sup} are the same,
and that all rows of df_sub appear in df_sup
}
