% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qens.R
\name{predict.qens}
\alias{predict.qens}
\title{Predict based on a quantile regression averaging fit}
\usage{
\method{predict}{qens}(qens_obj, qfm, sort_quantiles)
}
\arguments{
\item{qfm}{matrix of model forecasts of class QuantileForecastMatrix}

\item{sort_quantiles}{logical; if TRUE, the predictive quantiles are sorted
in order of the quantile level.  Otherwise, the raw predictive quantiles are
returned.}

\item{qens}{object of class qens}
}
\value{
data frame with ensemble quantile forecasts
}
\description{
Predict based on a quantile regression averaging fit
}
